Nouvelle fonction dans l'API REST et màj des tests et schema SQL. J'ai rajouté à l'API l'implémentation de la fonction getSubordinate avec la route /employee/subordinate/get/all.
Les tests ont subis quelques modification.
-test_create_root_session, après insertion les id sont récupérés pour mettre à
jour les tables Session et Person;
-test_personapi, meilleur gestion des exceptions pour empêcher le programme
de ne pas aller en fin d'exécution;
-test_persistence_api, test de la fonction SessionAPI::findByToken;

Modification de la table Person, ajout d'une colonne pour le clé étrangère de la
table Token. la longueur de la colonne passwd à 2048 caractères.

+++ include/API/SessionAPI.hpp
@@ -79,7 +79,7 @@ namespace Util
     {
         Entity::Session t_session;
         t_session.setBegin((ulong) std::time(nullptr)) ;
-        t_session.setEnd(t_session.getBegin()+31556952UL) ;
+        t_session.setEnd(t_session.getBegin()+(ulong)31556952) ;
         t_session.setPerson(p_person) ;
         t_session.setToken(hashSha512(p_person->getEmail())) ;
         return t_session ;
